@@ -6,10 +6,11 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-scripts": "^5.0.1",
     "semantic-ui-css": "^2.2.11",
+    "uuid": "^10.0.0",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
